remove dependency on ts-mockito and remove original MockUtils
