+++ test/unit-tests/ui/ReloadExtension.test.ts
@@ -14,7 +14,7 @@
 import { expect } from "chai";
 import { showReloadExtensionNotification } from "../../../src/ui/ReloadExtension";
 import * as vscode from "vscode";
-import { mockNamespace } from "../MockUtils2";
+import { mockNamespace } from "../MockUtils";
 
 suite("ReloadExtension Unit Test Suite", async function () {
     const windowMock = mockNamespace(vscode, "window");
